extract building events view
